refactor: split SitesTable into separate variant components (Phase 2 - part 4)
- Extract mobile accordion variant to SitesTable/SitesTableMobile.tsx (261 lines)
- Extract desktop variants to SitesTable/SitesTableDesktop.tsx (340 lines)
- Create router component SitesTable/index.tsx (50 lines, 90% reduction)
- Remove monolithic SitesTable.tsx (540 lines)
- Update test imports to new folder structure

Benefits:
- Clear separation of mobile vs desktop concerns
- Each variant independently maintainable
- Main entry point reduced from 540 to 50 lines
- Easier to test variants in isolation
- Follows Open/Closed Principle

All 184 tests passing
Linter clean (0 errors, 0 warnings)
App verified working in browser
